Added math library
